examples brought up-to-date

+++ examples/addNonIncludedFiles.pl
@@ -11,6 +11,7 @@ my $rootDir="$nvidiaSdk/C/src/";
 my @filePairs=(
 "simpleTextureDrv/simpleTexture_kernel.cu,texture.cu",
 "simpleTextureDrv/data/lena_bw.pgm,lena_bw.pgm",
+"simpleTextureDrv/data/ref_rotated.pgm,ref_rotated.pgm"
 );
 
 chdir("../");
@@ -18,5 +19,8 @@ chdir("../");
 
 foreach my $filePair (@filePairs) {
 	my @temp = split(/,/,$filePair);
+	next if (-r "examples/$temp[1]"); # don't overwrite
+	next unless (-r "$rootDir/$temp[0]"); # if source doesn't exist, continue
 	system("cp -i $rootDir/$temp[0] examples/$temp[1]");
 }
+
